python/requirements.txt
simplejson==3.13.2
pbr==3.1.1
paho-mqtt==1.3.1
cloudant==2.7.0
